requirements.txt
flask
flask-oauthlib
python-jose
requests
